re-organise tests and fixtures
