+++ tests/conftest.py
@@ -1,14 +1,7 @@
 """conftest.py."""
 
-from .fixtures.fixtures_dataframe import (  # noqa: F401
-    bng_index_column,
-    boolean_column,
-    bounds_column,
-    cell_size,
-    data_for_rasterisation,
-    dataframe_for_rasterisations,
+from .fixtures.dataframe.call_method import (  # noqa: F401
     dataframe_other,
-    expected_attrs,
     expected_dataframe,
     expected_dataframe_filter,
     expected_dataframe_grouped,
@@ -17,34 +10,40 @@ from .fixtures.fixtures_dataframe import (  # noqa: F401
     expected_dataframe_schema,
     expected_dataframe_select,
     expected_dataframewrapper_name,
-    expected_empty_metadata,
-    expected_geometry_mask,
-    expected_hl_dataset_no_metadata,
-    expected_hl_dataset_with_metadata,
     expected_schema_grouped,
     expected_schema_joined,
     expected_schema_select,
+    schema_other,
+    temp_path,
+)
+from .fixtures.dataframe.common import (  # noqa: F401
+    expected_empty_metadata,
+    spark_session,
+)
+from .fixtures.dataframe.to_zarr import (  # noqa: F401
+    bng_index_column,
+    boolean_column,
+    bounds_column,
+    cell_size,
+    data_for_rasterisation,
+    dataframe_for_rasterisations,
+    expected_float32_dataset,
+    expected_float64_dataset,
+    expected_int16_dataset,
+    expected_int32_dataset,
+    expected_int64_dataset,
+    expected_mask_dataset,
     float32_column,
     float64_column,
     geometry_column,
-    geometry_mask_path,
-    hl_dataframe,
-    hl_schema,
-    hl_wrapper_no_metadata,
-    hl_wrapper_with_metadata,
-    hl_zarr_path_no_metadata,
-    hl_zarr_path_with_metadata,
-    int8_column,
     int16_column,
     int32_column,
     int64_column,
+    make_expected_dataset_path,
     num_rows,
     out_shape,
     schema_for_rasterisation,
-    schema_other,
-    spark_session,
     string_category_column,
-    temp_path,
 )
 from .fixtures.fixtures_graph import expected_dag  # noqa: F401
 from .fixtures.fixtures_metadata import (  # noqa: F401
